requirements.txt
Django==5.0.3
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
mysqlclient==2.2.0
coverage==7.5.0